requirements.txt
matplotlib~=3.4.2
numpy~=1.18.5
Pillow~=7.2.0
torch~=1.8.1+cu101
torchvision~=0.9.1+cu101
tqdm~=4.59.0
wandb~=0.10.30
imageio~=2.9.0
opencv-python~=4.5.2.52
future~=0.18.2